+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/Vector256.cs
@@ -2386,13 +2386,13 @@ namespace System.Runtime.Intrinsics
             for (int index = 0; index < Vector256<sbyte>.Count; index++)
             {
                 byte selectedIndex = (byte)indices.GetElementUnsafe(index);
-                byte selectedValue = 0;
+                sbyte selectedValue = 0;
 
                 if (selectedIndex < Vector256<byte>.Count)
                 {
                     selectedValue = vector.GetElementUnsafe(selectedIndex);
                 }
-                result.SetElementUnsafe(index, (sbyte)selectedValue);
+                result.SetElementUnsafe(index, selectedValue);
             }
 
             return result;
